+++ nd/tests/test_utils.py
@@ -11,6 +11,7 @@ from nd.testing import (equal_list_of_dicts, generate_test_dataset,
 import time
 from datetime import datetime
 from dateutil.tz import tzutc
+from collections import OrderedDict
 
 
 @pytest.mark.parametrize('fmt', [
@@ -291,3 +292,59 @@ def test_is_complex():
 def test_is_complex_invalid_input():
     with assert_raises_regex(ValueError, 'Not an xarray Dataset or DataArray'):
         utils.is_complex('a string')
+
+
+@pytest.mark.parametrize('dims', [
+    ('x', 'y'), ('time',)
+])
+def test_apply(dims):
+    ds = generate_test_dataset()
+    ref = ds.mean(dims)
+    result = utils.apply(
+        ds, np.mean, signature='({})->()'.format(",".join(dims))
+    )
+    xr_assert_allclose(
+        result.transpose(*ref.nd.dims),
+        ref.transpose(*ref.nd.dims)
+    )
+
+
+def test_apply_with_vars():
+    ds = generate_test_dataset()
+    ref = ds.to_array(dim='var').mean('var')
+    result = utils.apply(
+        ds, lambda a: a.mean(axis=1), signature='(time,var)->(time)'
+    )
+    xr_assert_allclose(
+        result.transpose(*ref.nd.dims),
+        ref.transpose(*ref.nd.dims)
+    )
+
+
+def test_apply_with_vars_keep_vars():
+    ds = generate_test_dataset()
+    ref = ds.mean('time')
+    result = utils.apply(
+        ds, lambda a: a.mean(axis=0), signature='(time,var)->(var)'
+    )
+    xr_assert_allclose(
+        result.transpose(*ref.nd.dims),
+        ref.transpose(*ref.nd.dims)
+    )
+
+
+@pytest.mark.parametrize('args,kwargs', [
+    ((1, 2, 3), dict(c=4, d=5)),
+    ((1,), dict(b=2, d=3)),
+    ((1, 2, 3, 4, 5), dict()),
+    ((), dict(b=2, a=1)),
+])
+def test_extract_arguments(args, kwargs):
+    def fn(a, b, *args, c=None, **kwargs):
+        return OrderedDict(
+            a=a, b=b, args=args, c=c, kwargs=kwargs
+        )
+
+    bound = utils.extract_arguments(fn, args, kwargs)
+    actual = fn(*args, **kwargs)
+    assert_equal(bound, actual)